-	CommentActivity.java I switched one php script, do_post_comment.php to beta_do_post_comment.php This new script has an extra little added code that posts to the user_requests_tbl which we're slowly turning into a general notification table. NOTE: php script is on the server
-	FriendRequestItem.java
I overloaded the public FriendRequestItem(int, String, String, String)
function so that it would take an additional integer argument. Added
[private int type] which will hold the type of notification it is.

Also aded a public int getter [public int getType()] which returns the
type of the notification.

-	FriendRequestItemAdapter
Added an if/else ladder to the getView function which will generate a
message on the view based on what type of notification it is. 

-	MainMenuActivity.java
Switched one php script, get_friend_requests.php to
beta_get_friend_requests.php
The new script has an included line for getting an additional table item
from the user_requests table.

Worked on listview to make a notification invisible once clicked.

Used my overloaded FriendrequestItem function with the additional
integer on the doInBackground function

commented out the Toast text in the onPostExecute function due to
additional notification items now present on the notification bar

Added an if/else ladder to the startDialog function so that based on the
type of notification that it is, the dialog box will be custom tailored
to the notification. The comment notification, FOR NOW, takes a user to
their profile. Thats messy, but it's a good start. The rewards
notification is still a work in progress. 

Apologies if I've omitted anything.